Refactoring: Extract non datastore related code from TokenDatastore #3104
