+++ AlphaWallet/Tokens/Types/TokensDataStore.swift
@@ -29,51 +29,10 @@ class TokensDataStore {
         return SessionManager(configuration: configuration)
     }()
 
-    private lazy var getNameCoordinator: GetNameCoordinator = {
-        return GetNameCoordinator(forServer: server)
+    private lazy var tokenProvider: TokenProviderType = {
+        return TokenProvider(account: account, server: server)
     }()
 
-    private lazy var getSymbolCoordinator: GetSymbolCoordinator = {
-        return GetSymbolCoordinator(forServer: server)
-    }()
-
-    private lazy var getNativeCryptoCurrencyBalanceCoordinator: GetNativeCryptoCurrencyBalanceCoordinator = {
-        return GetNativeCryptoCurrencyBalanceCoordinator(forServer: server)
-    }()
-
-    private lazy var getERC20BalanceCoordinator: GetERC20BalanceCoordinator = {
-        return GetERC20BalanceCoordinator(forServer: server)
-    }()
-
-    private lazy var getERC875BalanceCoordinator: GetERC875BalanceCoordinator = {
-        return GetERC875BalanceCoordinator(forServer: server)
-    }()
-
-    private lazy var getERC721ForTicketsBalanceCoordinator: GetERC721ForTicketsBalanceCoordinator = {
-        return GetERC721ForTicketsBalanceCoordinator(forServer: server)
-    }()
-
-    private lazy var getIsERC875ContractCoordinator: GetIsERC875ContractCoordinator = {
-        return GetIsERC875ContractCoordinator(forServer: server)
-    }()
-
-    private lazy var getERC721BalanceCoordinator: GetERC721BalanceCoordinator = {
-        return GetERC721BalanceCoordinator(forServer: server)
-    }()
-
-    private lazy var getIsERC721ForTicketsContractCoordinator: GetIsERC721ForTicketsContractCoordinator = {
-        return GetIsERC721ForTicketsContractCoordinator(forServer: server)
-    }()
-
-    private lazy var getIsERC721ContractCoordinator: GetIsERC721ContractCoordinator = {
-        return GetIsERC721ContractCoordinator(forServer: server)
-    }()
-
-    private lazy var getDecimalsCoordinator: GetDecimalsCoordinator = {
-        return GetDecimalsCoordinator(forServer: server)
-    }()
-
-    private let account: Wallet
     private let assetDefinitionStore: AssetDefinitionStore
     private let realm: Realm
     private var pricesTimer = Timer()
@@ -91,7 +50,9 @@ class TokensDataStore {
     private let openSea: OpenSea
     private let queue = DispatchQueue.global()
 
+    let account: Wallet
     let server: RPCServer
+
     weak var delegate: TokensDataStoreDelegate?
     weak var priceDelegate: TokensDataStorePriceDelegate?
     weak var erc721TokenIdsFetcher: Erc721TokenIdsFetcher?
@@ -220,277 +181,9 @@ class TokensDataStore {
         refreshBalance()
     }
 
-    func getContractName(for address: AlphaWallet.Address,
-                         completion: @escaping (ResultResult<String, AnyError>.t) -> Void) {
-        withRetry(times: numberOfTimesToRetryFetchContractData) { [weak self] triggerRetry in
-            guard let strongSelf = self else { return }
-            strongSelf.getNameCoordinator.getName(for: address) { (result) in
-                switch result {
-                case .success:
-                    completion(result)
-                case .failure:
-                    if !triggerRetry() {
-                        completion(result)
-                    }
-                }
-            }
-        }
-    }
-
-    func getContractSymbol(for address: AlphaWallet.Address,
-                           completion: @escaping (ResultResult<String, AnyError>.t) -> Void) {
-        withRetry(times: numberOfTimesToRetryFetchContractData) { [weak self] triggerRetry in
-            guard let strongSelf = self else { return }
-            strongSelf.getSymbolCoordinator.getSymbol(for: address) { result in
-                switch result {
-                case .success:
-                    completion(result)
-                case .failure:
-                    if !triggerRetry() {
-                        completion(result)
-                    }
-                }
-            }
-        }
-    }
-
-    func getDecimals(for address: AlphaWallet.Address,
-                     completion: @escaping (ResultResult<UInt8, AnyError>.t) -> Void) {
-        withRetry(times: numberOfTimesToRetryFetchContractData) { [weak self] triggerRetry in
-            guard let strongSelf = self else { return }
-            strongSelf.getDecimalsCoordinator.getDecimals(for: address) { result in
-                switch result {
-                case .success:
-                    completion(result)
-                case .failure:
-                    if !triggerRetry() {
-                        completion(result)
-                    }
-                }
-            }
-        }
-    }
-
-    func getContractName(for address: AlphaWallet.Address) -> Promise<String> {
-        Promise { seal in
-            getContractName(for: address) { (result) in
-                switch result {
-                case .success(let name):
-                    seal.fulfill(name)
-                case .failure(let error):
-                    seal.reject(error)
-                }
-            }
-        }
-    }
-
-    func getContractSymbol(for address: AlphaWallet.Address) -> Promise<String> {
-        Promise { seal in
-            getContractSymbol(for: address) { result in
-                switch result {
-                case .success(let name):
-                    seal.fulfill(name)
-                case .failure(let error):
-                    seal.reject(error)
-                }
-            }
-        }
-    }
-
-    func getDecimals(for address: AlphaWallet.Address) -> Promise<UInt8> {
-        Promise { seal in
-            getDecimals(for: address) { result in
-                switch result {
-                case .success(let name):
-                    seal.fulfill(name)
-                case .failure(let error):
-                    seal.reject(error)
-                }
-            }
-        }
-    }
-
-    func getTokenType(for address: AlphaWallet.Address) -> Promise<TokenType> {
-        Promise { seal in
-            getTokenType(for: address) { tokenType in
-                seal.fulfill(tokenType)
-            }
-        }
-    }
-
-    func getERC20Balance(for address: AlphaWallet.Address, completion: @escaping (ResultResult<BigInt, AnyError>.t) -> Void) {
-        withRetry(times: numberOfTimesToRetryFetchContractData) { [weak self] triggerRetry in
-            guard let strongSelf = self else { return }
-            strongSelf.getERC20BalanceCoordinator.getBalance(for: strongSelf.account.address, contract: address) { result in
-                switch result {
-                case .success:
-                    completion(result)
-                case .failure:
-                    if !triggerRetry() {
-                        completion(result)
-                    }
-                }
-            }
-        }
-    }
-
-    func getERC875Balance(for address: AlphaWallet.Address,
-                          completion: @escaping (ResultResult<[String], AnyError>.t) -> Void) {
-        withRetry(times: numberOfTimesToRetryFetchContractData) { [weak self] triggerRetry in
-            guard let strongSelf = self else { return }
-            strongSelf.getERC875BalanceCoordinator.getERC875TokenBalance(for: strongSelf.account.address, contract: address) { result in
-                switch result {
-                case .success:
-                    completion(result)
-                case .failure:
-                    if !triggerRetry() {
-                        completion(result)
-                    }
-                }
-            }
-        }
-    }
-
-    func getERC721ForTicketsBalance(for address: AlphaWallet.Address,
-                                    completion: @escaping (ResultResult<[String], AnyError>.t) -> Void) {
-        withRetry(times: numberOfTimesToRetryFetchContractData) { [weak self] triggerRetry in
-            guard let strongSelf = self else { return }
-            strongSelf.getERC721ForTicketsBalanceCoordinator.getERC721ForTicketsTokenBalance(for: strongSelf.account.address, contract: address) { result in
-                switch result {
-                case .success:
-                    completion(result)
-                case .failure:
-                    if !triggerRetry() {
-                        completion(result)
-                    }
-                }
-            }
-        }
-    }
-
-    //TODO should callers call tokenURI and so on, instead?
-    func getERC721Balance(for address: AlphaWallet.Address, completion: @escaping (ResultResult<[String], AnyError>.t) -> Void) {
-        withRetry(times: numberOfTimesToRetryFetchContractData) { [weak self] triggerRetry in
-            guard let strongSelf = self else { return }
-            strongSelf.getERC721BalanceCoordinator.getERC721TokenBalance(for: strongSelf.account.address, contract: address) { result in
-                switch result {
-                case .success(let balance):
-                    if balance >= Int.max {
-                        completion(.failure(AnyError(Web3Error(description: ""))))
-                    } else {
-                        completion(.success([String](repeating: "0", count: Int(balance))))
-                    }
-                case .failure(let error):
-                    if !triggerRetry() {
-                        completion(.failure(error))
-                    }
-                }
-            }
-        }
-    }
-
     private func getTokensFromOpenSea() -> OpenSea.PromiseResult {
         //TODO when we no longer create multiple instances of TokensDataStore, we don't have to use singleton for OpenSea class. This was to avoid fetching multiple times from OpenSea concurrently
         return openSea.makeFetchPromise(forOwner: account.address)
-    }
-
-    func getTokenType(for address: AlphaWallet.Address, completion: @escaping (TokenType) -> Void) {
-        let isErc875Promise = Promise<Bool> { seal in
-            withRetry(times: numberOfTimesToRetryFetchContractData) { [weak self] triggerRetry in
-                guard let strongSelf = self else { return }
-                //Function hash is "0x4f452b9a". This might cause many "execution reverted" RPC errors
-                //TODO rewrite flow so we reduce checks for this as it causes too many "execution reverted" RPC errors and looks scary when we look in Charles proxy. Maybe check for ERC20 (via EIP165) as well as ERC721 in parallel first, then fallback to this ERC875 check
-                strongSelf.getIsERC875ContractCoordinator.getIsERC875Contract(for: address) { [weak self] result in
-                    guard self != nil else { return }
-                    switch result {
-                    case .success(let isERC875):
-                        if isERC875 {
-                            seal.fulfill(true)
-                        } else {
-                            seal.fulfill(false)
-                        }
-                    case .failure:
-                        if !triggerRetry() {
-                            seal.fulfill(false)
-                        }
-                    }
-                }
-            }
-        }
-        enum Erc721Type {
-            case erc721
-            case erc721ForTickets
-            case notErc721
-        }
-        let isErc721Promise = Promise<Erc721Type> { seal in
-            withRetry(times: numberOfTimesToRetryFetchContractData) { [weak self] triggerRetry in
-                guard let strongSelf = self else { return }
-                strongSelf.getIsERC721ContractCoordinator.getIsERC721Contract(for: address) { [weak self] result in
-                    guard let strongSelf = self else { return }
-                    switch result {
-                    case .success(let isERC721):
-                        if isERC721 {
-                            withRetry(times: strongSelf.numberOfTimesToRetryFetchContractData) { [weak self] triggerRetry2 in
-                                guard let strongSelf = self else { return }
-                                strongSelf.getIsERC721ForTicketsContractCoordinator.getIsERC721ForTicketContract(for: address) { result in
-                                    switch result {
-                                    case .success(let isERC721ForTickets):
-                                        if isERC721ForTickets {
-                                            seal.fulfill(.erc721ForTickets)
-                                        } else {
-                                            seal.fulfill(.erc721)
-                                        }
-                                    case .failure:
-                                        if !triggerRetry2() {
-                                            seal.fulfill(.erc721)
-                                        }
-                                    }
-                                }
-                            }
-                        } else {
-                            seal.fulfill(.notErc721)
-                        }
-                    case .failure:
-                        if !triggerRetry() {
-                            seal.fulfill(.notErc721)
-                        }
-                    }
-                }
-            }
-        }
-
-        firstly {
-            isErc721Promise
-        }.done { isErc721 in
-            switch isErc721 {
-            case .erc721:
-                completion(.erc721)
-            case .erc721ForTickets:
-                completion(.erc721ForTickets)
-            case .notErc721:
-                break
-            }
-        }.cauterize()
-
-        firstly {
-            isErc875Promise
-        }.done { isErc875 in
-            if isErc875 {
-                completion(.erc875)
-            } else {
-                //no-op
-            }
-        }.cauterize()
-
-        firstly {
-            when(fulfilled: isErc875Promise.asVoid(), isErc721Promise.asVoid())
-        }.done { _, _ in
-            if isErc875Promise.value == false && isErc721Promise.value == .notErc721 {
-                completion(.erc20)
-            } else {
-                //no-op
-            }
-        }.cauterize()
     }
 
     func tokenThreadSafe(forContract contract: AlphaWallet.Address) -> TokenObject? {
@@ -540,7 +233,7 @@ class TokensDataStore {
         case .nativeCryptocurrency:
             completion()
         case .erc20:
-            getERC20Balance(for: tokenObject.contractAddress, completion: { [weak self] result in
+            tokenProvider.getERC20Balance(for: tokenObject.contractAddress, completion: { [weak self] result in
                 defer { completion() }
                 guard let strongSelf = self else { return }
                 switch result {
@@ -551,7 +244,7 @@ class TokensDataStore {
                 }
             })
         case .erc875:
-            getERC875Balance(for: tokenObject.contractAddress, completion: { [weak self] result in
+            tokenProvider.getERC875Balance(for: tokenObject.contractAddress, completion: { [weak self] result in
                 defer { completion() }
                 guard let strongSelf = self else { return }
                 switch result {
@@ -564,7 +257,7 @@ class TokensDataStore {
         case .erc721:
             break
         case .erc721ForTickets:
-            getERC721ForTicketsBalance(for: tokenObject.contractAddress, completion: { [weak self] result in
+            tokenProvider.getERC721ForTicketsBalance(for: tokenObject.contractAddress, completion: { [weak self] result in
                 defer { completion() }
                 guard let strongSelf = self else { return }
                 switch result {
@@ -710,7 +403,7 @@ class TokensDataStore {
     }
 
     func refreshETHBalance() {
-        getNativeCryptoCurrencyBalanceCoordinator.getBalance(for: account.address) {  [weak self] result in
+        tokenProvider.getEthBalance(for: account.address) {  [weak self] result in
             guard let strongSelf = self else { return }
             switch result {
             case .success(let balance):
